Update README.md
This repository contains my C++ solutions to various competitive programming problems from platforms such as Codeforces and UVA judge, as well as exercises that I have completed for practice. In this repository, you will find a variety of algorithms and problem-solving techniques that I have applied to tackle challenges of varying difficulty. This collection serves as a resource for anyone wishing to improve their understanding of C++ in the context of competitive programming, and reflects my ongoing journey to improve my coding skills.